requirements.txt
slack-bolt>=1.18.0
anthropic>=0.40.0
python-dotenv>=1.0.0
streamlit>=1.35.0
